rerender keyed elements
Ensure that when we rerender, Elements with different keys are
considered different Elements. This is important for the case that a
View(layout="row") is replaced by a View(layout="column"), because that
layout prop cannot change, the entire widget must be destroyed and
recreated. Ideally the layout prop could change.