use_async_call: return canceller from hook call
By delaying the return of the cancel function until
the trigger function is called we're requiring the
user to store the canceller manually using the
`use_state` hook in many common scenarios. By returning
this function directly we can eliminate this extra complexity
at the expense of adding an _ variable for the canceller.