app: properly close with pending tasks
The logic for when to exit the application was to check whether the
_hook_async defaultdict was empty. However, any component that has
pending tasks while the component is deleted from the tree will never be
deleted from the _hook_async dictionary. Thus forcing the application to
never close.

This commit updates the logic to also check each of the remaining
components if they have finished, and if so, delete them from the
dictionary such that the program may exit cleanly.