Deprecate View, ScrollView
Replace View with:
- HBoxView
- VBoxView
- FixView

Replace ScrollView with:
- HScrollView
- VScrollView
- FixScrollView

We are eliminating the `layout= "row" | "column" | "none"` prop because it
never worked as a prop. If the value of the layout prop was
changed then the layout of the View would not change. To replace a
View(layout="row") with a View(layout="column"), it was necessary to add
unique set_key()s so that the reconciliation algorithm would recognize
that the Views needed to be destroyed and recreated.

ButtonView is now a subclass of HBoxView and has row layout instead of
column layout by default. For column layout, put an HBoxView inside the
ButtonView.